+++ eelbrain/_experiment/tests/test_sample_experiment.py
@@ -589,7 +589,7 @@ def test_sample_source(samples_experiment):
         two_stage_manifest_data = json.load(fid)
     assert 'two-stage-level-1' in {dep['name'] for dep in two_stage_manifest_data['dependencies'].values()}
     subject_dep = two_stage_manifest_data['dependencies']['R0000']
-    with open(Path(subject_dep['manifest'])) as fid:
+    with open(e._derivatives.cache_dir / subject_dep['manifest']) as fid:
         level_1_manifest_data = json.load(fid)
     assert level_1_manifest_data['dependencies']['two-stage-data']['dependencies']['R0000']['name'] == 'evoked-stc'
     ds_return, _ = e.load_test('twostage', 0.05, 0.2, 0.05, samples=8, return_data=True, make=True)